Clarified tests structure and added impact normal tests
