+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "shocovox-rs"
-version = "0.3.0"
+version = "0.3.1"
 edition = "2021"
 authors = ["Dávid Tóth <toth.david.munka@gmail.com>"]
 license = "MIT OR Apache-2.0"
