+++ src/spatial/math/tests.rs
@@ -1,0 +1,67 @@
+#[cfg(feature = "raytracing")]
+#[cfg(test)]
+mod intersection_tests {
+
+    use crate::spatial::{
+        math::{cube_impact_normal, plane_line_intersection},
+        raytracing::Ray,
+        Cube, V3c,
+    };
+
+    #[test]
+    fn test_negative_intersection() {
+        let plane_point = V3c::new(0., 0., 0.);
+        let plane_normal = V3c::new(0., 1., 0.);
+        let line_origin = V3c::new(0., 1., 0.);
+        let line_direction = V3c::new(0., 1., 0.);
+        assert!(plane_line_intersection(
+            &plane_point,
+            &plane_normal,
+            &line_origin,
+            &line_direction
+        )
+        .is_some_and(|v| v == -1.));
+    }
+
+    #[test]
+    fn test_edge_case_cube_top_hit() {
+        let ray = Ray {
+            origin: V3c {
+                x: 8.965594,
+                y: 10.0,
+                z: -4.4292345,
+            },
+            direction: V3c {
+                x: -0.5082971,
+                y: -0.72216684,
+                z: 0.46915793,
+            },
+        };
+        let t_hit = (Cube {
+            min_position: V3c::new(2.0, 0.0, 0.0),
+            size: 2.0,
+        })
+        .intersect_ray(&ray)
+        .unwrap();
+
+        assert!(t_hit
+            .impact_distance
+            .is_some_and(|v| (v - 11.077772).abs() < 0.001));
+        assert!((ray.point_at(t_hit.impact_distance.unwrap()).y - 2.).abs() < 0.001);
+    }
+
+    #[test]
+    fn test_impact_normal() {
+        let cube = Cube {
+            min_position: V3c::unit(0.),
+            size: 2.,
+        };
+
+        assert!(V3c::new(0., 0., 1.) == cube_impact_normal(&cube, &V3c::new(1., 1., 2.)));
+        assert!(V3c::new(0., 1., 0.) == cube_impact_normal(&cube, &V3c::new(1., 2., 1.)));
+        assert!(V3c::new(1., 0., 0.) == cube_impact_normal(&cube, &V3c::new(2., 1., 1.)));
+        assert!(V3c::new(0., 0., -1.) == cube_impact_normal(&cube, &V3c::new(1., 1., 0.)));
+        assert!(V3c::new(0., -1., 0.) == cube_impact_normal(&cube, &V3c::new(1., 0., 1.)));
+        assert!(V3c::new(-1., 0., 0.) == cube_impact_normal(&cube, &V3c::new(0., 1., 1.)));
+    }
+}
